Split models into shared/pro/nonpro
