+++ nanokvm/client.py
@@ -24,48 +24,28 @@ from PIL import Image
 from pydantic import BaseModel, ValidationError
 import yarl
 
-from .models import (
+from .models.common import (
     ApiResponse,
     ApiResponseCode,
     ChangePasswordReq,
     ConnectWifiReq,
-    DeleteEdidReq,
     DeleteImageReq,
-    DiskType,
     DownloadImageReq,
     GetAccountRsp,
-    GetCdRomRsp,
-    GetCustomEdidListRsp,
-    GetEdidRsp,
     GetGpioRsp,
     GetHardwareRsp,
-    GetHdmiCaptureRsp,
-    GetHdmiPassthroughRsp,
-    GetHdmiStateRsp,
     GetHidModeRsp,
     GetHostnameRsp,
     GetImagesRsp,
     GetInfoRsp,
-    GetKvmadminStatusRsp,
-    GetLcdTimeFormatRsp,
-    GetLedStripRsp,
-    GetLowPowerRsp,
     GetMdnsStateRsp,
-    GetMemoryLimitRsp,
-    GetMenuBarConfigRsp,
     GetMountedImageRsp,
     GetMouseJigglerRsp,
     GetOLEDRsp,
     GetPreviewRsp,
     GetSSHStateRsp,
-    GetStaticIPRsp,
-    GetSwapSizeRsp,
     GetTailscaleStatusRsp,
-    GetTimeStatusRsp,
-    GetTimeZoneRsp,
     GetVersionRsp,
-    GetVirtualDeviceProRsp,
-    GetVirtualDeviceRsp,
     GetWebTitleRsp,
     GetWifiRsp,
     GpioType,
@@ -79,36 +59,60 @@ from .models import (
     MouseButton,
     MouseJigglerMode,
     PasteReq,
+    SetGpioReq,
+    SetHidModeReq,
+    SetHostnameReq,
+    SetMouseJigglerReq,
+    SetOledReq,
+    SetPreviewReq,
+    SetWebTitleReq,
+    StatusImageRsp,
+    UpdateVirtualDeviceReq,
+    VirtualDevice,
+    WakeOnLANReq,
+)
+from .models.non_pro import (
+    GetCdRomRsp,
+    GetHdmiStateRsp,
+    GetMemoryLimitRsp,
+    GetSwapSizeRsp,
+    GetVirtualDeviceRsp,
+    SetMemoryLimitReq,
+    SetSwapSizeReq,
+)
+from .models.pro import (
+    DeleteEdidReq,
+    DiskType,
+    GetCustomEdidListRsp,
+    GetEdidRsp,
+    GetHdmiCaptureRsp,
+    GetHdmiPassthroughRsp,
+    GetKvmadminStatusRsp,
+    GetLcdTimeFormatRsp,
+    GetLedStripRsp,
+    GetLowPowerRsp,
+    GetMenuBarConfigRsp,
+    GetStaticIPRsp,
+    GetTimeStatusRsp,
+    GetTimeZoneRsp,
+    GetVirtualDeviceProRsp,
     RateControlMode,
     RefreshVirtualDeviceReq,
     ScanWifiRsp,
     SetFpsReq,
     SetGopReq,
-    SetGpioReq,
     SetHdmiCaptureReq,
     SetHdmiPassthroughReq,
-    SetHidModeReq,
-    SetHostnameReq,
     SetLcdTimeFormatReq,
     SetLedStripReq,
     SetLowPowerReq,
-    SetMemoryLimitReq,
     SetMenuBarConfigReq,
-    SetMouseJigglerReq,
-    SetOledReq,
-    SetPreviewReq,
     SetRateControlModeReq,
     SetStaticIPReq,
     SetStreamModeReq,
     SetStreamQualityReq,
-    SetSwapSizeReq,
     SetTimeZoneReq,
-    SetWebTitleReq,
-    StatusImageRsp,
     SwitchEdidReq,
-    UpdateVirtualDeviceReq,
-    VirtualDevice,
-    WakeOnLANReq,
 )
 from .utils import obfuscate_password
